fix memory leak in MICO SSL initialization
This patch fixes memory leak in MICO SSL initialization. The problem
was caused by the fact that MICOSSL::_init operation is called from
ORB_init and it did not check if it was already called or not.

darcs-hash:c740f6e1eb08e30dfdb78dc75c8ef2ce24c29d45